treams_core::lattice::sheets

Function self_sheet

pub(super) fn self_sheet(k: Complex, eta: Complex) -> f64

The sheet arg v = 2 arg k - pi - arg((k eta)^2) of v = -1 / (2 eta^2) for self_term: -pi continues the outgoing side Im v = -0 of real k and splits, and k takes arg k = pi on the negative real axis for either sign of zero, as the AMOS outgoing waves do.